Fixed issue when creating a shadow copy of a package, if the universe aligns together and the cosmos unit together to    produce a package length exactly to where the end-of-package byte sequence is cut in half, the shadow copy will fail    to be created due to the end-of-package byte sequence being cut in half, this issue was fixed by reading the package    in chunks to determine the end-of-package byte sequence.
The PackageReader class has been updated to use a more efficient buffer management when reading data from a file. Previously, buffer size was growing uncontrollably and can eventually lead to out-of-memory errors for large files.

Now, the data is read in chunks and the buffer is cleared when it exceeds approximately 1MB size, maintaining only the last 512KB. This change ensures a more memory-efficient package reading and effectively prevents erroneous deadlocks for large package files.

Additionally, detection for end-of-data byte sequence has been modified to rectify an issue where package length could cut off the end-of-package byte sequence. This results in an improved reliability for package validation.